requirements.txt
a2wsgi==1.7.0
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==40.0.1
ecdsa==0.18.0
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
PyMySQL==1.0.3
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0

starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
websockets==11.0.1
fastapi-utils==0.2.1

